requirements.txt
Babel==2.14.0
eagerpy==0.30.0
frozenlist==1.4.1
gudhi==3.9.0
imageio==2.22.2
jaxtyping==0.2.25
matplotlib==3.6.1
matplotlib-inline==0.1.6
monai==1.0.0
networkx==2.8.7
nibabel==5.2.0
numpy
Pillow==9.2.0
pybind11==2.11.1
pybind11-global==2.11.1
PyYAML==6.0
scikit-image==0.20.0
scikit-learn==1.3.2
scipy==1.9.2
sklearn==0.0
tqdm==4.64.1
wandb==0.16.3
yarg==0.1.9
yarl==1.9.4
